+++ cli/abeamer-cli.ts
@@ -367,9 +367,6 @@ The commands are:
                 + '\n\n' + all;
             });
           }
-
-          // readjusts file paths
-          text = text.replace(/\.\.\/\.\.\/client\/lib/g, 'abeamer');
           break;
       }